requirements.txt
# Core dependencies
google-genai>=1.18.0
sentence-transformers>=4.1.0
faiss-cpu>=1.11.0
PyPDF2>=3.0.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0

# Environment and utilities
python-dotenv>=1.0.0
typing-extensions>=4.0.0

# Optional: For GPU support (replace faiss-cpu with faiss-gpu)
# faiss-gpu>=1.11.0

# Optional: For additional embedding models
# scikit-learn>=1.3.0
# scipy>=1.10.0 